Treat null tokens in mech strings as numerically zero
This removes the special case for "standard" clan mechs, as that was
doing the same thing in a sepcific place; this handles a more general
case. With this change the preference for putting numbers first means
that if two strings share a common prefix, the one with fewer tokens
will always be sorted first.

For example, this causes "Rifleman C" to sort before "Rifleman C 2".
This breaks the model strings down to ["C", 0, 0] vs ["C", "", 2];
the preference for numbers before strings means the comparison
terminates at the second token, putting the 0 before "". Previously the
two would compare equal for the middle token because " " is normalized
to "", making it equal to the empty string extensions added into the
shorter token list.